+++ images/webrtc-stream/webrtc/js/localmedia.js
